storewebapi/storeapi/storeapi/requirements.txt
Django==4.0
djangorestframework==3.13.1
numpy==1.21.5
pandas==1.3.5
